Added ray intersection for spheres. Changed return type of Shape.IntersectRay to double.

+++ src/RayTracing/Shape.java
@@ -3,5 +3,5 @@ package RayTracing;
 
 public interface Shape
 {
-    float IntersectRay(Ray ray);
+    double IntersectRay(Ray ray);
 }
